requirements.txt
Jinja2
json
einops 
pillow
opencv-python
timm transformers
tqdm
pandas
torch
transformers
pandas
numpy
easyocr